refactor(ui): move the filter search bar into @everr/ui
The bar is a plain input with a hint and a clear button. Nothing in it knows
about telemetry, and all four explorers reach it through the package name
already, so it sits with the other primitives.

packages/ui gains the DOM test setup its sibling has, jsdom plus
testing-library, so the component keeps its test. Its vitest config aliases
the package name onto src, the way the components import each other.
