+++ packages/ui/package.json
@@ -28,8 +28,12 @@
     "zod": "catalog:"
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.2",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
-    "typescript": "catalog:"
+    "jsdom": "^29.1.1",
+    "typescript": "catalog:",
+    "vitest": "^4.1.5"
   }
 }
